+++ spec/logidze/model_spec.rb
@@ -578,4 +578,30 @@ describe Logidze::Model, :db do
       it { is_expected.to be_zero }
     end
   end
+
+  describe '.reset_log_data' do
+    let!(:user2) { user.dup.tap(&:save!) }
+    let!(:user3) { user.dup.tap(&:save!) }
+
+    before { User.limit(2).reset_log_data }
+
+    it 'nullify log_data column for a association' do
+      expect(user.reload.log_size).to be_zero
+      expect(user2.reload.log_size).to be_zero
+    end
+
+    it 'not affect other model records' do
+      expect(user3.reload.log_size).to eq 5
+    end
+  end
+
+  describe '#reset_log_data' do
+    subject { user.log_size }
+
+    before { user.reset_log_data }
+
+    it 'nullify log_data column for a single record' do
+      is_expected.to be_zero
+    end
+  end
 end